chanserv/help: include INFO in short help instead of RECOVER
The RECOVER command should rarely be needed unless handing out channel
access like popcorn is the norm in some channel, in which case we can't
really help them anyway; as such there's little point in listing it by
default. INFO on the other hand is widely used as *the* way to ask
services for basic information on a channel, so it makes sense to be in
the short help (which nickserv/help does as well).

(In fact, on freenode, the most common use of RECOVER seems to be by
people who use it as the primary way to get opped, given that network's
recommendation to avoid auto-op and their default flag set reflecting
this. I suppose some people misinterpret this as having lost control of
their channel. Presumably it's debatable who's to blame here, but maybe
removing RECOVER from the default short help listing will help avoid
that sort of problem. Either way, it shouldn't hurt.)

(cherry picked from commit 4396f465e9f7fd993432d84e275a38703d215aea)