nickserv/multimark: fix patricia warning
The code to deal with nick/account (un)registrations operated in the
following manner:

- retrieve the restored mark list for a name (creating one if there
  isn't already)
- modify this list
- try to add the updated restored mark list for the used name

However, the list is already modified; the patricia needn't be updated.
In fact, the mowgli_patricia_add call was useless as it merely refused
to add a duplicate key,

(patricia.c:748 mowgli_patricia_elem_add): Key is already in dict, ignoring duplicate

There may be further room for improvement of the code for a future
release.