crypto/argon2d: simplify code
Division modulo 1 is 0, so ref_lane would have always been 0,
which means same_lane is always true, which means argon2d_idx()
can be dramatically simplified: ra_size will always be at least
(ctx->index - 1), so start with that and add more as needed,
bearing in mind that it can't be decremented now, as it is only
ever decremented when same_lane is false.

This is an overlooked remnant of the threading logic of the
algorithm, which was ripped out at the start.